api/requirements.txt
# ============================
# API framework
# ============================
fastapi>=0.128,<1
uvicorn[standard]>=0.23,<1
python-multipart>=0.0.22,<1
requests>=2.32,<3

# ============================
# Image & numerical processing
# ============================
numpy>=1.26,<3
pillow>=10.3,<11
matplotlib>=3.10,<4
opencv-python-headless>=4.13,<5

# ============================
# Deep learning
# ============================
tensorflow==2.20.0
keras==3.12.0
h5py>=3.15,<4

# ============================
# BLIP / Transformers
# ============================
torch>=2.10,<3
torchvision>=0.25,<1
transformers>=4.52,<5
sentencepiece>=0.1.99
huggingface-hub>=0.36,<1
safetensors>=0.4,<1
